requirements.txt
Django==2.1.1
django-heroku==0.3.1
django-storages==1.6.3
gunicorn==19.6.0
whitenoise==4.1
boto3==1.9.12
sorl-thumbnail==12.5.0
pillow==5.2.0
django-randomslugfield==0.3.0
django-markdownx==2.0.24
